+++ Entities/Concrete/Car.cs
@@ -12,7 +12,7 @@ namespace Entities.Concrete
         public int CarId { get; set; }
         public int BrandId { get; set; }
         public int ColorId { get; set; }
-        public int ModelYear { get; set; }
+        public short ModelYear { get; set; }
         public string Model { get; set; }
         public string Description { get; set; }
         public decimal DailyPrice { get; set; }
